refactor:create note component
